Add new documentation sections.
Reformat the navigation scheme, now driven by _config.yml

Make the topbar highlight the right section. This is done by having
all posts have a 'section' entry in their metadata. This section is
then matched with the sections in the topbar defined in _config. If
they match, the section is highlighted in the top bar.

Clean the CSS files that aren't used. Make sure they're all in /css